Add dependent name flags
Symbols with a name that was created using other symbols can not be used as part of name of other symbols. This is to prevent side effects. I might be able to come up with a solution for this in the future, but this is good enough for now. More investigation is required.