Refactor compiler pipeline
Expanding is just a part of what needs to happen after situation. This group of operations will include tabulation, prototyping, evaluating, resolving, and specializing. the thing all these operations have in common is that they require context information, and previous stages of compiler do not. so name it the "contextualization" stage.