Refactor argument and parameter situations
Replace binding and field situations with argument and parameter situations. Situator no longer detects certain kinds of errors such as signature having  positional parameters after named. this will be detected in a later compiler stage. this is better because it will allow for more clear log error messages, and it makes things a lot simpler.